+++ src/front-door/azext_front_door/tests/latest/test_waf_scenarios.py
@@ -11,6 +11,22 @@ from azure.core.exceptions import (HttpResponseError)
 
 
 class WafTests(WafScenarioMixin, ScenarioTest):
+    @live_only()  # --defer seems not work with VCR.py well
+    @ResourceGroupPreparer(location='westus')
+    def test_waf_log_scrubbing(self, resource_group):
+        blockpolicy = self.create_random_name(prefix='cli', length=24)
+        cmd = 'az network front-door waf-policy create -g {resource_group} -n {blockpolicy} --mode prevention --sku Standard_AzureFrontDoor'.format(**locals())
+        result = self.cmd(cmd).get_output_in_json()
+        self.assertEqual(result['name'], blockpolicy)
+        self.assertEqual(result['policySettings']['mode'], "Prevention")
+        self.assertEqual(result['policySettings']['requestBodyCheck'], "Enabled")
+        self.assertIn('customRules', result)
+        self.assertIn('managedRules', result)
+        self.assertIn('id', result)
+        options = '--log-scrubbing \"{{scrubbing-rules:[{{match-variable:QueryStringArgNames,selector-match-operator:EqualsAny}}],state:Enabled}}\"'
+        cmd = 'az network front-door waf-policy update -g {resource_group} -n {blockpolicy}'.format(**locals())
+        result = self.cmd(cmd + ' ' + options).get_output_in_json()
+        self.assertEqual(result['policySettings']['logScrubbing']['state'], "Enabled")
 
     @live_only()  # --defer seems not work with VCR.py well
     @ResourceGroupPreparer(location='westus')
@@ -105,9 +121,6 @@ az network front-door waf-policy delete -g {resource_group} -n {policyName}
         self.assertEqual(result['name'], detectiondisabledpolicy)
         self.assertEqual(result['policySettings']['mode'], "Detection")
         self.assertEqual(result['policySettings']['enabledState'], "Disabled")
-        self.assertEqual(result['policySettings']['redirectUrl'], None)
-        self.assertEqual(result['policySettings']['customBlockResponseStatusCode'], None)
-        self.assertEqual(result['policySettings']['customBlockResponseBody'], None)
         self.assertEqual(result['policySettings']['requestBodyCheck'], "Enabled")
         self.assertIn('customRules', result)
         self.assertIn('managedRules', result)